src/bsmartflow-acc-bridge/requirements.txt
Flask
google-auth
requests
gunicorn
